feat: add foot module, screening appointment system and dashboard improvements

+++ backend/src/main/java/com/basic/backend/entity/ScreeningAppointment.java
@@ -12,7 +12,7 @@ public class ScreeningAppointment {
 
     private Long userId;
 
-    private Integer screeningTypeId;
+    private Long screeningTypeId;
 
     private LocalDate scheduledFor;
